move other functions that call skipInternalf to ginkgo_wrappers
